Cleanup of Resharper warnings

+++ Api/0.0.0.1/EtAlii.Ubigia.Api.Logical/Content/Queries/ContentPartQueryHandler.cs
@@ -27,7 +27,7 @@
         {
             var contentPart = await _fabric.Content.Retrieve(identifier, contentPartId);
             var buffer = contentPart.Data;
-            localDataStream.Write(buffer, 0, (int)buffer.Length);
+            localDataStream.Write(buffer, 0, buffer.Length);
         }
     }
 }
